perf(durable-streams-rust): serve caught-up long-polls from the epoll reactor
Generalize the SSE reactor (renamed sse_reactor -> reactor) to serve two reader
modes from one slab/epoll/flush/wake core. A caught-up long-poll on a reactor-
eligible stream now hands its socket to the reactor instead of parking the whole
connection-task future for the wait; the reactor writes the single response on
data/close/timeout and, for a keep-alive client, hands the socket back to the
async connection loop so the client's next request is read and dispatched there
(re-registering with the reactor if it is another caught-up poll). Reusing the
existing async keep-alive path keeps the reactor free of a second HTTP request
reader. Linux only; non-Linux and cold/forked/tiered streams keep the inline
wait. SSE behaviour is unchanged.
